fixed some things to make it work
for self-host

+++ src/models/issue.cr
@@ -11,6 +11,6 @@ class Issue < BaseModel
     belongs_to author : User
     belongs_to assignee : User?
     column status : Issue::AvramStatus
-    column scoped_id : Int32
+    column scoped_id : Int64
   end
 end
